default alarm sound changed, multiple alarm sounds can be used, howler alarm GUI (SSP, partially SMP), alarmRange now used for industrial alarm and as default value of howler alarm

+++ common/net/minecraft/src/nuclearcontrol/IC2NuclearControl.java
@@ -30,7 +30,7 @@ public abstract class IC2NuclearControl extends NetworkMod implements IGuiHandle
     public static Item itemRangeUpgrade;
     public static Block blockNuclearControlMain;
     public static int modelId;
-    public static float alarmRange;
+    public static int alarmRange;
     protected static IC2NuclearControl instance;
